requirements.txt
Django
psycopg2
psycopg2-binary
python-decouple
pillow
django-role-permissions
djangorestframework